+++ CHANGELOG.md
@@ -4,6 +4,12 @@
  - `pio update` has a new api: `pio update app <options>`, `pio update ui <options>`.
  - added version of UI & firmware to `pio version -v`
  - PioreactorUI has a different way to update, rather than using git. We now version the PioreactorUI, so it's easier to know if which version is being used.
+ - `pio update app` now has a `--version` option to specify a version of the Pioreactor software.
+ - correctly publish `alt_media_fraction` to MQTT in dosing jobs
+
+
+### 22.12.3
+ - fix for chemostat
 
 ### 22.12.2
  - Remove some errant debugging statements